Fix nuget fallback volume mount for dotnet debugging
Previously the NuGetFallbackFolder mounted into the container used for
debugging was hardcoded to /usr/local/share/dotnet/sdk/NuGetFallbackFolder
for all non-windows platforms. However this path is only the valid
default for OSX. On Linux /usr/share/dotnet/sdk/NuGetFallbackFolder is
the default installation location. This patch corrects that behavior.