Always Add host and port to `compose.debug.yml`
It appears that we are looking for an `npm` script which directly references `node` in order to generate the proper debug configuration (--inspect=0.0.0.0:9229). But directly referencing `node` isn't required for an `npm start` script. For instance, the following is valid...

```
"scripts": {
  "start": "./server.js"
}
```

I had this configuration and it took me quite a while to figure out why the debugger wouldn't attach.

It seems to me that we should always generate the proper host:port configuration for the `debug.yml`. What is the use case for not doing that if we don't find "node" as a command in the `package.json` scripts section? If the user has already selected "node", then node is implicit.