Fix "webview is disposed" error (#2959)
* Fix "webview is disposed" error

Fixes issue https://github.com/microsoft/vscode-docker/issues/2820

The poblem is likely due to a handful of async calls done beteween when the start page view is created and when its content is set. That gives the user an opportunity to close the start page before we attempt to set its content. It is especially tempting since during the time the content is computed the already-visible start page panel is empty.

This change front-loads that async processing so that the webview creation becomes fully synchronous.