Merge rel/1.13.0 to main (#2970)
* Prepare for release (#2929)

* Upgrade the vsce to 1.88 to get latest underscore (#2930)

* Add extendedLocation property in Plan and Site creation (#2931)

* Add extendedLocation property in Plan and Site creation

* Updating comment

* Ravipal/merge from main (#2964)

* Add malware scan (#2942)

* Add malware scan.

* Actually add malware scan.

* Move malware task and stop immediately.

* Update UI dependency for new icon. (#2955)

* Update vsce dependency (#2957)

* Fix "webview is disposed" error (#2959)

* Fix "webview is disposed" error

Fixes issue https://github.com/microsoft/vscode-docker/issues/2820

The poblem is likely due to a handful of async calls done beteween when the start page view is created and when its content is set. That gives the user an opportunity to close the start page before we attempt to set its content. It is especially tempting since during the time the content is computed the already-visible start page panel is empty.

This change front-loads that async processing so that the webview creation becomes fully synchronous.

* Add missing description for debug config property. (#2963)

Co-authored-by: Phillip Hoff <phillip@orst.edu>
Co-authored-by: Karol Zadora-Przylecki <karolz@microsoft.com>

* Update change log to includ LIMA support (#2965)

* Update date on release notes (#2966)

Co-authored-by: Phillip Hoff <phillip@orst.edu>
Co-authored-by: Karol Zadora-Przylecki <karolz@microsoft.com>