Update Dockerfile language server to 0.11.0
This update adds support for a few new flags that have been added to
Dockerfile so that they no longer get flagged as errors. The other big
change is the introduction of a comment that can be inserted into the
Dockerfile to instruct the linter to not flag something as an error.
While most errors can be ignored this way, clearly fatal errors that
suggest the Dockerfile is completely broken cannot be ignored.

Signed-off-by: Remy Suen <remy.suen@gmail.com>